Move most modules to modules/
