+++ src/edge_probing.py
@@ -7,7 +7,7 @@ import torch.nn.functional as F
 from torch.autograd import Variable
 
 from .tasks.tasks import EdgeProbingTask
-from . import modules
+from .modules import modules
 
 from allennlp.modules.span_extractors import \
     EndpointSpanExtractor, SelfAttentiveSpanExtractor
